% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cj_euclid}
\alias{cj_euclid}
\title{Fit generalized Euclidean preferences model}
\usage{
cj_euclid(
  formula,
  data,
  X = labels(terms(formula))[[1]],
  Y = labels(terms(formula))[[2]],
  Row = NULL,
  Col = NULL,
  mins = NULL,
  maxs = NULL,
  lengths = NULL,
  fixed_effects = NULL,
  ideal_color = "blue",
  x_vals = c("L", "M", "H"),
  y_vals = c("A", "B", "C"),
  x_breaks = NULL,
  y_breaks = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula}

\item{data}{A \code{data.frame}.}

\item{X}{string: variable on X axis for plot}

\item{Y}{string: variable on Y axis for plot}

\item{Row}{variable name for faceting by row}

\item{Col}{variable name for faceting by column}

\item{mins}{minimal values for fitting variables}

\item{maxs}{minimal values for fitting variables}

\item{lengths}{lengths of fitting variables}

\item{ideal_color}{color for ideal points}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Fit generalized Euclidean preferences model
}
\examples{
library(tidyverse)

out <-
  lm_euclid(
   rating ~ universal + stringency + severity,
   data = covid_policy_evaluations)
out <-
  cj_euclid(
   rating ~ universal + stringency + severity,
   data = covid_policy_evaluations,
   fixed_effects = "ID",
   lengths = c(20, 20, 3),
   Col = "severity")
out$graph
}
