% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cjEuclid}
\alias{cjEuclid}
\title{Fit generalized Euclidean preferences model}
\usage{
cjEuclid(
  data,
  Y = "Y",
  A = "A",
  B = "B",
  C = "C",
  ID = "ID",
  a_min = -1,
  a_max = 1,
  a_step = 1,
  b_min = -1,
  b_max = 1,
  b_step = 0.1,
  c_min = -1,
  c_max = 1,
  c_step = 0.1,
  ideal_color = "blue",
  x_vals = c("L", "M", "H"),
  y_vals = c("A", "B", "C"),
  x_breaks = c(-0.9, 0, 0.9),
  y_breaks = c(-0.9, 0, 0.9)
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{A, }{B, C, Y names of policy and outcome variables}

\item{ideal_color}{Color for idea points}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}
}
\value{
A  \code{list}
}
\description{
Fit generalized Euclidean preferences model
}
\examples{
library(tidyverse)
library(DeclareDesign)
data <-
fabricatr::fabricate(
  ID = add_level(20),
  choice = add_level(2,
                     A = rnorm(N), B = rnorm(N), C = rnorm(N),
                     Y = -(A^2 + (B-.3)^2 + (C --.67)^2) + .2*rnorm(N)) + A*B)
out <- cjEuclid(data)
}
\keyword{internal}
