% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_policy_evaluations}
\alias{covid_policy_evaluations}
\title{Covid freedoms conjoint data}
\format{
A data frame with 41480  rows and 12 columns:
\describe{
  \item{ID}{Anonymized unique subject ID}
  \item{round}{First or second round}
  \item{vignette}{First or second displayed vignette}
  \item{vaccinated}{Binary: whether vaccinated or not}
  \item{party.id}{}
  \item{severity}{}
  \item{stringency}{}
  \item{universal}{}
  \item{rating}{}
  \item{choice}{}
  \item{trust}{}
  \item{vaccination_probability}{Probability of getting vaccinated}
}
}
\usage{
data(covid_policy_evaluations)
}
\description{
Data from "Trading Liberties: Estimating policy ideal points and indifference curves from conjoint data”
Felix Hartmann, Macartan Humphreys, Ferdinand Geissler, Heike Klüver, and Johannes Giesecke
}
\references{
Hartman et al. (2023)
}
\keyword{datasets}
