% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{euclid_fits}
\alias{euclid_fits}
\title{Get fitted values}
\usage{
euclid_fits(
  formula,
  model,
  data,
  fixed_effects = NULL,
  mins = NULL,
  maxs = NULL,
  lengths = NULL
)
}
\arguments{
\item{formula}{A formula indicating lhs and rhs vars, e.g. Y ~ A + B + C}

\item{data}{A \code{data.frame}.}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}

\item{ideal_color}{Color for idea points}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Get fitted values
}
\examples{
model <- lm_euclid(rating ~ universal + stringency,
   data = covid_policy_evaluations,
   fixed_effects = "ID")$model

predictions_df <-
  euclid_fits(
    formula = rating ~ universal + stringency,
    model = model,
    data = covid_policy_evaluations,
    fixed_effects = "ID",
    lengths = c(5, 5))

}
