% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{euclid_plot}
\alias{euclid_plot}
\title{Plot fitted values with ideal points marked}
\usage{
euclid_plot(
  predictions_df,
  X,
  Y,
  outcome = "utility",
  R = NULL,
  C = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  y_vals = c("L", "M", "H"),
  x_vals = c("l", "c", "r"),
  ideal_color = "blue"
)
}
\arguments{
\item{X}{string: variable on X axis for plot}

\item{Y}{string: variable on Y axis for plot}

\item{R}{variable name for faceting by row}

\item{C}{variable name for faceting by column}

\item{y_vals}{labels for y axis}

\item{x_vals}{labels for x axis}

\item{ideal_color}{color for ideal points}

\item{perdictions_df}{Predictions}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Plot fitted values with ideal points marked
}
\examples{
model <- lm_euclid(Y ~ A + B + C, data)
predictions_df <- euclid_fits(formula = Y ~ A + B + C, model, data, lengths = c(3,3, 3))
predictions_df <- euclid_fits(data, model, formula = Y ~ A + B, lengths = c(3,3))
 euclid_plot(predictions_df, X = "A", Y = "B")
 euclid_plot(predictions_df, X = "A", Y = "B", R = "C")
}
