% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cjEuclid}
\alias{cjEuclid}
\title{Fit generalized Euclidean preferences model}
\usage{
cjEuclid(
  formula,
  data,
  X = labels(terms(formula))[[1]],
  Y = labels(terms(formula))[[2]],
  R = NULL,
  C = NULL,
  mins = NULL,
  maxs = NULL,
  lengths = c(20, 20),
  ideal_color = "blue",
  x_vals = c("L", "M", "H"),
  y_vals = c("A", "B", "C"),
  x_breaks = NULL,
  y_breaks = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula}

\item{data}{A \code{data.frame}.}

\item{X}{string: variable on X axis for plot}

\item{Y}{string: variable on Y axis for plot}

\item{R}{variable name for faceting by row}

\item{C}{variable name for faceting by column}

\item{ideal_color}{color for ideal points}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Fit generalized Euclidean preferences model
}
\examples{
library(tidyverse)
library(DeclareDesign)
data <-
fabricatr::fabricate(
  ID = add_level(20),
  choice = add_level(2,
                     A = rnorm(N), B = rnorm(N), C = rnorm(N),
                     Y = -(A^2 + (B-.3)^2 + (C --.67)^2) + .2*rnorm(N)))
out <- cjEuclid(Y ~ A + B + C, data, lengths = c(20, 20, 3))
out$graph
}
