% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{euclid_fits}
\alias{euclid_fits}
\title{Get fitted values}
\usage{
euclid_fits(formula, model, data, mins = NULL, maxs = NULL, lengths = NULL)
}
\arguments{
\item{formula}{A formula indicating lhs and rhs vars, e.g. Y ~ A + B + C}

\item{data}{A \code{data.frame}.}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}

\item{ideal_color}{Color for idea points}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Get fitted values
}
\examples{
model <- lm_euclid(Y ~ A + B + C, data)
predictions_df <- euclid_fits(formula = Y ~ A + B + C, model, data, lengths = c(3,3, 3))
}
