% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{euclid_plot}
\alias{euclid_plot}
\title{Plot fitted values with ideal points marked}
\usage{
euclid_plot(
  predictions_df,
  X,
  Y,
  outcome = "utility",
  Row = NULL,
  Col = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  y_vals = c("L", "M", "H"),
  x_vals = c("l", "c", "r"),
  ideal_color = "blue"
)
}
\arguments{
\item{X}{string: variable on X axis for plot}

\item{Y}{string: variable on Y axis for plot}

\item{Row}{variable name for faceting by row}

\item{Col}{variable name for faceting by column}

\item{y_vals}{labels for y axis}

\item{x_vals}{labels for x axis}

\item{ideal_color}{color for ideal points}

\item{perdictions_df}{Predictions}

\item{...}{other arguments passed to estimatr}
}
\value{
A  \code{list}
}
\description{
Plot fitted values with ideal points marked
}
\examples{

model <- lm_euclid(rating ~ universal + stringency,
   data = covid_policy_evaluations,
   fixed_effects = ~ ID, se_type = "stata")$model

predictions_df <-
  euclid_fits(
    formula = rating ~ universal + stringency,
    model = model,
    fixed_effects = "ID",
    data = covid_policy_evaluations,
    lengths = c(50, 50))

 euclid_plot(predictions_df, "universal", "stringency")
}
