% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lm_euclid}
\alias{lm_euclid}
\title{Fit Euclidean model}
\usage{
lm_euclid(formula, data, ...)
}
\arguments{
\item{formula}{A formula indicating lhs and rhs vars, e.g. Y ~ A + B + C}

\item{data}{A \code{data.frame}.}

\item{...}{other arguments passed to estimatr}

\item{x_vals}{labels for x axis}

\item{y_vals}{labels for y axis}

\item{ideal_color}{Color for idea points}
}
\value{
A  \code{list}
}
\description{
Fit Euclidean model
}
\examples{
library(tidyverse)
library(DeclareDesign)
data <-
fabricatr::fabricate(
  ID = add_level(50),
  choice = add_level(2,
                     A = rnorm(N), B = rnorm(N), C = rnorm(N),
                     Y = -(A^2 + (B-.3)^2 + (C --.67)^2) + .1*rnorm(N)))
lm_euclid(Y ~ A + B, data, fixed_effects = ~ ID)
}
